see version 2.4.0 changelog

+++ selenium_driver_updater/util/extractor.py
@@ -13,13 +13,13 @@ sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.
 class Extractor():
 
     @staticmethod
-    def extract_all_zip_archive(archive_path : str, out_path : str) -> Tuple[bool, str]:
+    def extract_all_zip_archive(archive_path : str, out_path : str, delete_archive : bool = True) -> Tuple[bool, str]:
         """Extract all members in specific zip archive
 
         Args:
             archive_path (str)  : Path to specific archive.
             out_path (str)      : Out path, where all members of archive will be gathered.
-
+            delete_archive (bool)   : Delete archive after unzip or not. Defaults to True.
 
         Returns:
             Tuple of bool and str
@@ -34,7 +34,7 @@ class Extractor():
             with zipfile.ZipFile(archive_path, 'r') as zip_ref:
                 zip_ref.extractall(out_path)
 
-            if os.path.exists(archive_path):
+            if os.path.exists(archive_path) and delete_archive:
                 os.remove(archive_path)
 
             result_run = True
@@ -47,12 +47,13 @@ class Extractor():
         return result_run, message_run
     
     @staticmethod
-    def extract_all_tar_gz_archive(archive_path : str, out_path : str) -> Tuple[bool, str]:
+    def extract_all_tar_gz_archive(archive_path : str, out_path : str, delete_archive : bool = True) -> Tuple[bool, str]:
         """Extract all members in specific tar.gz archive
 
         Args:
             archive_path (str)  : Path to specific archive.
             out_path (str)      : Out path, where all members of archive will be gathered.
+            delete_archive (bool)   : Delete archive after unzip or not. Defaults to True.
 
         Returns:
             Tuple of bool and str
@@ -67,7 +68,7 @@ class Extractor():
             with tarfile.open(archive_path, "r:gz") as tar_ref:
                 tar_ref.extractall(out_path)
 
-            if os.path.exists(archive_path):
+            if os.path.exists(archive_path) and delete_archive:
                 os.remove(archive_path)
 
             result_run = True
@@ -79,7 +80,7 @@ class Extractor():
         return result_run, message_run
 
     @staticmethod
-    def extract_all_zip_archive_with_specific_name(archive_path : str, out_path : str, filename : str, filename_replace : str) -> Tuple[bool, str]:
+    def extract_all_zip_archive_with_specific_name(archive_path : str, out_path : str, filename : str, filename_replace : str, delete_archive : bool = True) -> Tuple[bool, str]:
         """Extract all zip archive and replaces name for one of member
 
         Args:
@@ -87,6 +88,7 @@ class Extractor():
             out_path (str)          : Out path, where all members of archive will be gathered.
             filename (str)          : Archive member whose name will be changed.
             filename_replace (str)  : Specific name for replacing.
+            delete_archive (bool)   : Delete archive after unzip or not. Defaults to True.
 
         Returns:
             Tuple of bool and str
@@ -106,13 +108,13 @@ class Extractor():
 
             if archive_path.endswith('.tar.gz'):
 
-                result, message = Extractor.extract_all_tar_gz_archive(archive_path=archive_path, out_path=driver_folder_path)
+                result, message = Extractor.extract_all_tar_gz_archive(archive_path=archive_path, out_path=driver_folder_path, delete_archive=delete_archive)
                 if not result:
                     return result, message
 
             elif archive_path.endswith('.zip'):
 
-                result, message = Extractor.extract_all_zip_archive(archive_path=archive_path, out_path=driver_folder_path)
+                result, message = Extractor.extract_all_zip_archive(archive_path=archive_path, out_path=driver_folder_path, delete_archive=delete_archive)
                 if not result:
                     return result, message
